fix: run lint:fix to fix formatting and auto-fixable lint issues
Agent-Logs-Url: https://github.com/WhiskeySockets/Baileys/sessions/9529f842-623e-4d7f-b414-e3819e8604ac

+++ src/Utils/chat-utils.ts
@@ -369,7 +369,7 @@ export const extractSyncdPatches = async (result: BinaryNode, options: RequestIn
 						content = Buffer.from(Object.values(content))
 					}
 
-					const syncd = proto.SyncdPatch.decode(content as Uint8Array)
+					const syncd = proto.SyncdPatch.decode(content)
 					if (!syncd.version) {
 						syncd.version = { version: +collectionNode.attrs.version! + 1 }
 					}
